Revert "Hosting MCP (#197)"
This reverts commit b72405c6b96697dd02184d8502c8e369930095f0.
