+++ .github/workflows/deploy-api.yaml
@@ -29,13 +29,8 @@ jobs:
             SUPABASE_SERVER_TOKEN
             TURSO_ORGANIZATION
             TURSO_GROUP_DATABASE_TOKEN
-            CF_ACCOUNT_ID
-            CF_API_TOKEN
         env:
           SUPABASE_URL: ${{ secrets.SUPABASE_URL }}
           SUPABASE_SERVER_TOKEN: ${{ secrets.SUPABASE_SERVER_TOKEN }}
           TURSO_ORGANIZATION: ${{ secrets.TURSO_ORGANIZATION }}
           TURSO_GROUP_DATABASE_TOKEN: ${{ secrets.TURSO_GROUP_DATABASE_TOKEN }}
-          CF_API_TOKEN: ${{ secrets.CF_API_TOKEN }}
-          CF_ACCOUNT_ID: ${{ secrets.CF_ACCOUNT_ID }}
-          CF_DISPATCH_NAMESPACE: "deco-chat-prod"
