--- .vscode/extensions.json
@@ -1,5 +1,0 @@
-{
-  "recommendations": [
-    "denoland.vscode-deno"
-  ]
-}
