+++ apps/api/main.ts
@@ -1,1 +1,5 @@
-export { default } from "./src/app.ts";
+import app from "./src/app.ts";
+
+export default {
+  fetch: app.fetch,
+};
