+++ apps/api/src/app.ts
@@ -1,43 +1,169 @@
-import { Hono } from "hono";
-import api from "./api.ts";
-import apps from "./apps.ts";
-import { AppEnv } from "./utils/context.ts";
+import { HttpServerTransport } from "@deco/mcp/http";
+import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
+import { Context, Hono } from "hono";
+import { cors } from "hono/cors";
+import { HTTPException } from "hono/http-exception";
+import { logger } from "hono/logger";
+import * as agentsAPI from "./api/agents/api.ts";
+import * as integrationsAPI from "./api/integrations/api.ts";
+import * as membersAPI from "./api/members/api.ts";
+import * as profilesAPI from "./api/profiles/api.ts";
+import * as teamsAPI from "./api/teams/api.ts";
+import * as threadsAPI from "./api/threads/api.ts";
+import { withContextMiddleware } from "./middlewares/context.ts";
+import { setUserMiddleware } from "./middlewares/user.ts";
+import { ApiHandler, createAIHandler, State } from "./utils/context.ts";
+import { ROUTES as loginRoutes } from "./auth/index.ts";
 
-const Hosts = {
-  API: "api.deco.chat",
-  APPS: "deco.page",
-} as const;
+const app = new Hono();
 
-export const APPS_DOMAIN_QS = "app_host";
-export const appsDomainOf = (req: Request, url?: URL) => {
-  url ??= new URL(req.url);
-  const referer = req.headers.get("referer");
+// Register tools for each API handler
+const GLOBAL_TOOLS = [
+  teamsAPI.getTeam,
+  teamsAPI.createTeam,
+  teamsAPI.updateTeam,
+  teamsAPI.deleteTeam,
+  teamsAPI.listTeams,
+  membersAPI.getTeamMembers,
+  membersAPI.addTeamMember,
+  membersAPI.updateTeamMember,
+  membersAPI.removeTeamMember,
+  profilesAPI.getProfile,
+  profilesAPI.updateProfile,
+];
 
-  return url.searchParams.get(APPS_DOMAIN_QS) ||
-    (referer && new URL(referer).searchParams.get(APPS_DOMAIN_QS));
-};
+// Tools tied to an specific workspace
+const WORKSPACE_TOOLS = [
+  agentsAPI.getAgent,
+  agentsAPI.deleteAgent,
+  agentsAPI.createAgent,
+  agentsAPI.updateAgent,
+  agentsAPI.listAgents,
+  integrationsAPI.getIntegration,
+  integrationsAPI.createIntegration,
+  integrationsAPI.updateIntegration,
+  integrationsAPI.deleteIntegration,
+  integrationsAPI.listIntegrations,
+  threadsAPI.listThreads,
+  threadsAPI.getThread,
+  threadsAPI.getThreadMessages,
+  threadsAPI.getThreadTools,
+];
 
-const normalizeHost = (req: Request) => {
-  const host = req.headers.get("host") ?? "localhost";
-  const appsHost = appsDomainOf(req);
-  if (appsHost) {
-    return Hosts.APPS;
+/**
+ * Creates and sets up an MCP server for the given tools
+ */
+const createMCPHandlerFor = (
+  tools: typeof GLOBAL_TOOLS | typeof WORKSPACE_TOOLS,
+) => {
+  const server = new McpServer(
+    { name: "@deco/api", version: "1.0.0" },
+    { capabilities: { tools: {} } },
+  );
+
+  for (const tool of tools) {
+    server.tool(
+      tool.name,
+      tool.description,
+      tool.schema.shape,
+      createAIHandler(tool.handler),
+    );
   }
-  return {
-    [Hosts.API]: Hosts.API,
-    localhost: Hosts.API,
-    "localhost:3001": Hosts.API,
-    "localhost:8000": Hosts.API,
-  }[host] ?? Hosts.APPS;
+
+  return async (c: Context) => {
+    const transport = new HttpServerTransport();
+
+    await server.connect(transport);
+
+    c.res = await State.run(
+      c,
+      transport.handleMessage.bind(transport),
+      c.req.raw,
+    );
+
+    return c.res;
+  };
 };
 
-export const app = new Hono<AppEnv>({
-  getPath: (req) =>
-    "/" +
-    normalizeHost(req) +
-    req.url.replace(/^https?:\/\/[^/]+(\/[^?]*)(?:\?.*)?$/, "$1"),
+/**
+ * Setup a handler for handling tool calls. It's used so that
+ * UIs can call the tools without suffering the serialization
+ * of the protocol.
+ */
+const createToolCallHandlerFor =
+  (tools: ApiHandler[]) => async (c: Context) => {
+    const tool = c.req.param("tool");
+    const args = await c.req.json();
+
+    const t = tools.find((t) => t.name === tool);
+
+    if (!t) {
+      throw new HTTPException(404, { message: "Tool not found" });
+    }
+
+    const { data, error } = t.schema.safeParse(args);
+
+    if (error || !data) {
+      throw new HTTPException(400, {
+        message: error?.message ?? "Invalid arguments",
+      });
+    }
+
+    const result = await State.run(c, t.handler, data);
+
+    return c.json({ data: result });
+  };
+
+// Add logger middleware
+app.use(logger());
+
+// Enable CORS for all routes on api.deco.chat and localhost
+app.use(cors({
+  origin: (origin) => origin,
+  allowMethods: ["HEAD", "GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS"],
+  allowHeaders: ["Content-Type", "Authorization", "Cookie", "Accept"],
+  exposeHeaders: ["Content-Type", "Authorization", "Set-Cookie"],
+  credentials: true,
+}));
+
+app.use(withContextMiddleware);
+app.use(setUserMiddleware);
+
+// MCP endpoint handlers
+app.all(
+  "/mcp",
+  createMCPHandlerFor(GLOBAL_TOOLS),
+);
+app.all(
+  "/:root/:slug/mcp",
+  createMCPHandlerFor(WORKSPACE_TOOLS),
+);
+
+// Tool call endpoint handlers
+app.post(
+  "/tools/call/:tool",
+  createToolCallHandlerFor(GLOBAL_TOOLS),
+);
+app.post(
+  "/:root/:slug/tools/call/:tool",
+  createToolCallHandlerFor(WORKSPACE_TOOLS),
+);
+
+// Login and auth routes
+Object.entries(loginRoutes).forEach(([route, honoApp]) => {
+  app.route(route, honoApp);
 });
 
-app.route(`/${Hosts.API}`, api);
-app.route(`/${Hosts.APPS}`, apps);
+// Health check endpoint
+app.get("/health", (c: Context) => c.json({ status: "ok" }));
+
+app.onError((err, c) => {
+  console.error(err);
+
+  return c.json(
+    { error: err?.message ?? "Internal server error" },
+    err instanceof HTTPException ? err.status : 500,
+  );
+});
+
 export default app;
