+++ apps/api/wrangler.toml
@@ -5,7 +5,6 @@ main = "main.ts"
 compatibility_date = "2024-11-27"
 routes = [
   { pattern = "api.deco.chat", custom_domain = true },
-  { pattern = "*/*", zone_name = "deco.page" },
 ]
 
 [vars]
@@ -14,6 +13,5 @@ VITE_USE_LOCAL_BACKEND = false
 [observability]
 enabled = true
 
-[[dispatch_namespaces]]
-binding = "PROD_DISPATCHER"
-namespace = "deco-chat-prod"
+# [[tail_consumers]]
+# service = "hyperdx-tail"
